+++ app/Http/Controllers/IndructionController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\IndructionRequest;
 use App\Http\Resources\ShowIndructionResource;
 use App\Models\Indruction;
 use Illuminate\Http\Request;
@@ -21,7 +22,7 @@ class IndructionController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(IndructionRequest $request)
     {
         $indruction = Indruction::indruction($request);
         return response()->json(['message'=>true, 'data'=>$indruction], 201);
@@ -34,7 +35,6 @@ class IndructionController extends Controller
     {
         $indruction = Indruction::find($id);
         if(!$indruction){
-
             return response()->json(['message'=>'Not found'],404);
         }
         $indruction = new ShowIndructionResource($indruction);
@@ -44,7 +44,7 @@ class IndructionController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(IndructionRequest $request, string $id)
     {
         $indruction = Indruction::indruction($request,$id);
         return response()->json(['Update indruction success'=>true, 'data'=>$indruction],200);
@@ -55,8 +55,10 @@ class IndructionController extends Controller
      */
     public function destroy(string $id)
     {
-        
         $indruction = Indruction::find($id);
+        if(!$indruction){
+            return response()->json(['message'=>'Not found'],404);
+        }
         $indruction->delete();
         return response()->json(['message'=>true, 'data'=>$indruction], 200);
     }
